chore(mobility): plan + PR1 — extract @klorad/secrets
MOBILITY_PLAN.md
  Step 0 deliverable per the build brief. Reuse / extract / net-new
  tables, a 15-PR sequence, two resolved decisions:

    1. Tenant model: reuse Project with a strict split.
       sceneData.mobility holds config only; devices / sources /
       status live in their own relation tables (FK on Project.id);
       credentials encrypted via @klorad/secrets, never in JSON.
    2. URL shape: mobility.klorad.com/<token>, mirrors Campus.

  Co-design rule for the eventual @klorad/sync extraction: must be
  shaped against both consumers (Campus Discovered + Mobility device
  sync), not just Campus, so it doesn't carry web-crawler
  assumptions into Mobility.

@klorad/secrets (PR 1 of the arc)
  Move AES-256-GCM at-rest secret helpers from
  apps/campus/lib/secrets.ts into a shared workspace package so
  every Klorad vertical can reuse them. Pure node:crypto utility —
  no DB, no env validation beyond the SECRETS_KEY read.

  apps/campus/lib/secrets.ts becomes a one-line re-export so the
  three existing import sites (assistant route, ai-settings, ai-
  settings/test) keep resolving without an edit. Zero behavioural
  change for Campus.

Branch convention per memory pr-workflow-one-branch:
feature/mobility-vertical is the single arc branch. Progressive
commits per PR phase. PR opens when the whole vertical is ready,
or earlier if reviewer asks.
