+++ apps/campus/package.json
@@ -24,6 +24,7 @@
     "@klorad/crawler": "workspace:*",
     "@klorad/design-system": "workspace:^",
     "@klorad/prisma": "workspace:*",
+    "@klorad/secrets": "workspace:*",
     "@klorad/storage": "workspace:^",
     "@klorad/ui": "workspace:^",
     "@mappedin/mappedin-js": "^6.19.0",
